refactor: optimize date and relative date format
- add DateFormat.swift extension for DateFormat and RelativeDateFormat
- add date and relative date format to ResultsViewManager
- consolidate date and relative date format on AnnotationsOutlineDataSource
